# api.quiz/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz Platform API",
        "description": "API documentation for Quiz Platform with User and Role management",
        "contact": {
            "email": "admin@quizplatform.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.quiz.test",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/quizzes": {
            "get": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "List quizzes",
                "operationId": "c104ed69adb2fcc838e14859bee82cb2",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": []
            },
            "post": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Create quiz",
                "operationId": "17442658f5bcce78980f375fe66c83f1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Quiz"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/quizzes/{id}": {
            "get": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Get quiz",
                "operationId": "ce361b465ca18f66045654d0ed977db8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": []
            },
            "put": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Update quiz",
                "operationId": "7a420580b24a0a69cc55a676e172f09d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Quiz"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Delete quiz",
                "operationId": "61dcdd818514499d9fbc968c684e17e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Quiz": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Permission"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "owner_id": {
                                "type": "integer"
                            },
                            "category_id": {
                                "type": "integer",
                                "nullable": true
                            },
                            "title": {
                                "type": "string"
                            },
                            "description": {
                                "type": "string"
                            },
                            "difficulty": {
                                "type": "string",
                                "enum": [
                                    "easy",
                                    "medium",
                                    "hard"
                                ]
                            },
                            "is_paid": {
                                "type": "boolean"
                            },
                            "price_cents": {
                                "type": "integer",
                                "nullable": true
                            },
                            "timer_seconds": {
                                "type": "integer",
                                "nullable": true
                            },
                            "randomize_questions": {
                                "type": "boolean"
                            },
                            "randomize_answers": {
                                "type": "boolean"
                            },
                            "allow_multiple_attempts": {
                                "type": "boolean"
                            },
                            "max_attempts": {
                                "type": "integer",
                                "nullable": true
                            },
                            "visibility": {
                                "type": "string",
                                "enum": [
                                    "public",
                                    "private"
                                ]
                            },
                            "status": {
                                "type": "string",
                                "enum": [
                                    "draft",
                                    "published"
                                ]
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "Permission": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "view users"
                    },
                    "guard_name": {
                        "type": "string",
                        "example": "web"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}